Переместил код в messaging и operations
В процессе ревью там мы отказались от понятия clientId в kafka.
Также вернули 1в1 сопоставление kafka message в operations и в messaging.
Внесли partition в settings.
