--- Messaging.Transports.Kafka/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="2GIS.NuClear.Messaging.API" version="2.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Metamodeling" version="2.1.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.API" version="2.0.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Utils" version="1.1.13" targetFramework="net461" />
-  <package id="Confluent.Kafka" version="0.11.0" targetFramework="net461" />
-  <package id="librdkafka.redist" version="0.11.2-custom" targetFramework="net461" />
-</packages>
