+++ NuClear.River.ValidationRules.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{C1C7C6AC-4AAC-45E4-94AC-DEC47A72F447}"
 	ProjectSection(SolutionItems) = preProject
@@ -65,8 +65,6 @@ Project("{FF286327-C783-4F7A-AB73-9BCBAD0D4460}") = "ValidationRules.Replication
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ValidationRules.Replication.DatabaseComparison", "Tests\ValidationRules.Replication.DatabaseComparison\ValidationRules.Replication.DatabaseComparison.csproj", "{3E0D8363-B4B7-475A-975E-6A2FDAB8831E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Messaging.Transports.Kafka", "Messaging.Transports.Kafka\Messaging.Transports.Kafka.csproj", "{1A28D9CC-C476-4D87-AABE-587041C18F92}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ValidationRules.Hosting.Common", "ValidationRules.Hosting.Common\ValidationRules.Hosting.Common.csproj", "{876CF585-D099-4C5A-949E-E76DA08764D8}"
 EndProject
 Global
@@ -127,10 +125,6 @@ Global
 		{3E0D8363-B4B7-475A-975E-6A2FDAB8831E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3E0D8363-B4B7-475A-975E-6A2FDAB8831E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3E0D8363-B4B7-475A-975E-6A2FDAB8831E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A28D9CC-C476-4D87-AABE-587041C18F92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A28D9CC-C476-4D87-AABE-587041C18F92}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A28D9CC-C476-4D87-AABE-587041C18F92}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1A28D9CC-C476-4D87-AABE-587041C18F92}.Release|Any CPU.Build.0 = Release|Any CPU
 		{876CF585-D099-4C5A-949E-E76DA08764D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{876CF585-D099-4C5A-949E-E76DA08764D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{876CF585-D099-4C5A-949E-E76DA08764D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
