+++ Tests/ValidationRules.Replication.Comparison.Tests/app.config
@@ -24,7 +24,7 @@
         <behaviors>
             <endpointBehaviors>
                 <behavior name="webEndpoint">
-                    <webHttp/>
+                    <webHttp />
                 </behavior>
             </endpointBehaviors>
         </behaviors>
@@ -34,4 +34,16 @@
             <endpoint address="https://validation21.api.test.erm.2gis.ru" binding="webHttpBinding" bindingConfiguration="WebHttpBinding_IRiverRestService" behaviorConfiguration="webEndpoint" contract="ValidationRules.Replication.Comparison.Tests.RiverService.IRiverRestService" name="River" />
         </client>
     </system.serviceModel>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-10.0.0.0" newVersion="10.0.0.0" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.ServiceBus" publicKeyToken="31bf3856ad364e35" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-2.1.0.0" newVersion="2.1.0.0" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
 </configuration>
