+++ Tests/ValidationRules.Replication.StateInitialization.Tests/packages.config
@@ -6,11 +6,11 @@
   <package id="2GIS.NuClear.Model.Common" version="2.0.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Storage.API" version="2.0.3-fix-collection-q0001" targetFramework="net461" />
+  <package id="2GIS.NuClear.Storage.API" version="2.1.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Utils" version="1.1.13" targetFramework="net461" />
   <package id="CommonServiceLocator" version="1.3" targetFramework="net461" />
   <package id="linq2db" version="1.8.3" targetFramework="net461" />
-  <package id="NuClear.River.StateInitialization" version="5.2.2-accessor-types-p0042" targetFramework="net461" />
+  <package id="NuClear.River.StateInitialization" version="7.0.0" targetFramework="net461" />
   <package id="NUnit" version="2.6.4" targetFramework="net461" />
   <package id="Unity" version="4.0.1" targetFramework="net461" />
   <package id="Unofficial.Microsoft.SQLServer.SMO.2014" version="12.0.2000.8" targetFramework="net461" />
