+++ ValidationRules.Hosting.Common/packages.config
@@ -1,15 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="2GIS.NuClear.Messaging.API" version="2.2.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.API" version="2.3.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.Transports.Kafka" version="2.3.3" targetFramework="net461" />
   <package id="2GIS.NuClear.Metamodeling" version="2.1.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Model.Common" version="2.0.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Storage.API" version="2.0.3-fix-collection-q0001" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.API" version="2.0.1" targetFramework="net461" />
+  <package id="2GIS.NuClear.Storage.API" version="2.1.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.API" version="2.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Utils" version="1.1.13" targetFramework="net461" />
-  <package id="Confluent.Kafka" version="0.11.0" targetFramework="net461" />
+  <package id="Confluent.Kafka.StrongName" version="0.11.0" targetFramework="net461" />
   <package id="librdkafka.redist" version="0.11.2-custom" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
-  <package id="NuClear.River.Hosting.Common" version="5.2.2-accessor-types-p0042" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
+  <package id="NuClear.River.Hosting.Common" version="7.0.0" targetFramework="net461" />
 </packages>
