+++ ValidationRules.OperationsProcessing/Transports/Kafka/IAmsSettingsFactory.cs
@@ -1,0 +1,10 @@
+﻿using NuClear.Messaging.API.Flows;
+using NuClear.Messaging.Transports.Kafka;
+
+namespace NuClear.ValidationRules.OperationsProcessing.Transports.Kafka
+{
+    public interface IAmsSettingsFactory
+    {
+        IKafkaMessageFlowReceiverSettings CreateReceiverSettings(IMessageFlow messageFlow);
+    }
+}
