+++ ValidationRules.Querying.Host/packages.config
@@ -4,19 +4,18 @@
   <package id="2GIS.NuClear.Settings" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Telemetry" version="0.13.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.API" version="2.0.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.Environment" version="2.0.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.Log4Net" version="2.0.1" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.API" version="2.2.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.Environment" version="2.2.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.Log4Net" version="2.2.0" targetFramework="net461" />
   <package id="CommonServiceLocator" version="1.3" targetFramework="net461" />
   <package id="Confluent.Kafka" version="0.11.0" targetFramework="net461" />
   <package id="librdkafka.redist" version="0.11.2-custom" targetFramework="net461" />
   <package id="linq2db" version="1.8.3" targetFramework="net461" />
-  <package id="log4net" version="2.0.3" targetFramework="net461" />
-  <package id="log4net.Ext.Json.1.2.13" version="1.2.13.20151101" targetFramework="net461" />
+  <package id="log4net" version="2.0.8" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
   <package id="Unity" version="4.0.1" targetFramework="net461" />
 </packages>
