+++ ValidationRules.StateInitialization.Host/packages.config
@@ -2,40 +2,41 @@
 <packages>
   <package id="2GIS.NuClear.Assembling" version="1.0.4" targetFramework="net461" />
   <package id="2GIS.NuClear.IdentityService.Client" version="1.0.5" targetFramework="net461" />
-  <package id="2GIS.NuClear.Messaging.API" version="2.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Messaging.Transports.CorporateBus" version="2.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Messaging.Transports.ServiceBus" version="2.2.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.API" version="2.3.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.Transports.CorporateBus" version="2.3.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.Transports.Kafka" version="2.3.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Messaging.Transports.ServiceBus" version="2.3.3" targetFramework="net461" />
   <package id="2GIS.NuClear.Metamodeling" version="2.1.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Model.Common" version="2.0.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsLogging.API" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsLogging.Transports.ServiceBus" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsLogging.Transports.SQLStore" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsProcessing.API" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsProcessing.Transports.ServiceBus" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsProcessing.Transports.SQLStore" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsTracking" version="5.3.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.OperationsTracking.API" version="5.3.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsLogging.API" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsLogging.Transports.ServiceBus" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsLogging.Transports.SQLStore" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsProcessing.API" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsProcessing.Transports.ServiceBus" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsProcessing.Transports.SQLStore" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsTracking" version="5.5.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.OperationsTracking.API" version="5.5.3" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Storage.API" version="2.0.3-fix-collection-q0001" targetFramework="net461" />
-  <package id="2GIS.NuClear.Storage.LinqToDB" version="2.0.3-fix-collection-q0001" targetFramework="net461" />
+  <package id="2GIS.NuClear.Storage.API" version="2.1.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Storage.LinqToDB" version="2.1.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Telemetry" version="0.13.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.API" version="2.0.1" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.Environment" version="2.0.1" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.API" version="2.2.0" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.Environment" version="2.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Utils" version="1.1.13" targetFramework="net461" />
   <package id="2GIS.NuClear.Utils.Legacy" version="1.1.13" targetFramework="net461" />
   <package id="2GIS.NuClear.WCF.Client" version="1.0.12" targetFramework="net461" />
-  <package id="Confluent.Kafka" version="0.11.0" targetFramework="net461" />
+  <package id="Confluent.Kafka.StrongName" version="0.11.0" targetFramework="net461" />
   <package id="Dapper.StrongName" version="1.50.2" targetFramework="net461" />
   <package id="EnterpriseLibrary.TransientFaultHandling" version="6.0.1304.0" targetFramework="net461" />
   <package id="EnterpriseLibrary.TransientFaultHandling.ServiceBus" version="6.0.1304.0" targetFramework="net461" />
   <package id="librdkafka.redist" version="0.11.2-custom" targetFramework="net461" />
   <package id="linq2db" version="1.8.3" targetFramework="net461" />
   <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.1" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
-  <package id="NuClear.River.Hosting.Common" version="5.2.2-accessor-types-p0042" targetFramework="net461" />
-  <package id="NuClear.River.Replication" version="5.2.2-accessor-types-p0042" targetFramework="net461" />
-  <package id="NuClear.River.StateInitialization" version="5.2.2-accessor-types-p0042" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
+  <package id="NuClear.River.Hosting.Common" version="7.0.0" targetFramework="net461" />
+  <package id="NuClear.River.Replication" version="7.0.0" targetFramework="net461" />
+  <package id="NuClear.River.StateInitialization" version="7.0.0" targetFramework="net461" />
   <package id="protobuf-net" version="2.0.0.669" targetFramework="net461" />
   <package id="ServiceBus.v1_1" version="1.0.6" targetFramework="net461" />
   <package id="Unofficial.Microsoft.SQLServer.SMO.2014" version="12.0.2000.8" targetFramework="net461" />
